requirements.txt
colorama==0.4.6
pdf2image==1.17.0
pillow==11.0.0
tqdm==4.66.6
